+++ src/card/CardType/components/CardTypeDropdown.tsx
@@ -1,5 +1,5 @@
 import styles from "./CardTypeDropdown.module.css";
-import { CARD_TYPE_LIST } from "@/pages/addCard/constants";
+import { CARD_TYPE_LIST } from "@/card/CardType/constants";
 import Dropdown from "@/components/Dropdown/Dropdown";
 import { CardType } from "@/pages/addCard/types";
