+++ src/pages/addCard/components/AddCardForm/components/AddCardForm.module.css
@@ -1,5 +1,5 @@
 .form {
   display: flex;
   flex-direction: column;
-  gap: 1.5rem;
+  gap: 2rem;
 }
