+++ src/pages/addCard/types/index.ts
@@ -1,4 +1,4 @@
-import { CARD_TYPE_LIST, CARD_TYPE_COLOR } from "../constants";
+import { CARD_TYPE_LIST, CARD_TYPE_COLOR } from "@card/CardType/constants";
 
 export type CardType = (typeof CARD_TYPE_LIST)[number];
 export type CardTypeColor = (typeof CARD_TYPE_COLOR)[CardType];
